requirements.txt
Django==1.10.1
django-filter==0.14.0
djangorestframework==3.4.6
mysqlclient==1.3.7
